Выносит createOffersSectionTemplateTemplate из утилит в темлпейт
